requirements.txt
PySide6==6.9.2
requests==2.32.3
google-genai>=1.46.0
httpx[socks]>=0.27